requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
nltk
gensim
wordcloud
openpyxl
torch
huggingface
transformers
wandb
pyyaml
gradio